requirements.txt
Flask>=3.0
opencv-python
opencv-contrib-python
numpy
mysql-connector-python